Software Engineering Daily???
Hi - You beat me to the punch!  I was going to email my 3-4 favorite podcasts to the Clojure mailing list, and then I see this!  Good idea!

So, not sure if you've seen the Software Engineering Daily podcast or not, but I think it is one of the best. Not 100% FP oriented, but very important topics to any FP developer.

Alan